-implemented voxel cascades -replaced hardcoded voxel cascade resolution/size values with constants -added support for spec consts as array sizes in VKPipelineCache -index of visualized voxel cascade can now be changed at runtime by pressing 1-3
Former-commit-id: a28b0eee0733afea3988588190898227c4448353

+++ VEngine/Code/Graphics/Vulkan/Pass/BuildIndexBufferPass.cpp
@@ -84,7 +84,7 @@ void VEngine::BuildIndexBufferPass::addToGraph(RenderGraph &graph, const Data &d
 
 	BufferViewHandle indirectDrawCmdBufferViewHandle = *data.m_indirectDrawCmdBufferViewHandle;
 	BufferViewHandle filteredIndicesBufferViewHandle = *data.m_filteredIndicesBufferViewHandle;
-	const uint32_t clusterListSize = clusterInfoList.size();
+	const uint32_t clusterListSize = static_cast<uint32_t>(clusterInfoList.size());
 
 	ResourceUsageDescription passUsages[]
 	{
